% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{api_get}
\alias{api_get}
\title{Make an API call to the Kenya Health Information System (KHIS) server}
\usage{
api_get(url_path, d2_session, ..., retry = 1, verbosity = 0)
}
\arguments{
\item{url_path}{The path to make an api call}

\item{d2_session}{the d2Session object, default is "d2_default_session",}

\item{...}{Name-value pairs that provide query parameters. Each value must be
either a length-1 atomic vector (which is automatically escaped) or NULL (which
is silently dropped).}

\item{retry}{number of times to try in case of failure,.}

\item{verbosity}{How much information to print?
- 0: no output
- 1: show headers
- 2: show headers and bodies
- 3: show headers, bodies and curl status message}
}
\value{
Parsed JSON
}
\description{
\code{api_get} function makes an API call to the KHIS server
}
