% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthCred-class.R
\name{AuthCred-class}
\alias{AuthCred-class}
\alias{AuthCred}
\title{Authorization Credential}
\description{
Authorization Credential

Authorization Credential
}
\details{
An \code{AuthCred} object manages the authorization credentials that make request
to the KHIS API server.

An \code{AuthCred} should be created through the constructor function
\code{\link[=init_AuthCred]{init_AuthCred()}}, which has more details on the arguments

For more details on the parameters, see \code{\link[=init_AuthCred]{init_AuthCred()}}
}
\keyword{internal}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config_path}}{Path to a JSON configuration file.}

\item{\code{username}}{The KHIS username.}

\item{\code{password}}{The KHIS password.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-AuthCred-new}{\code{AuthCred$new()}}
\item \href{#method-AuthCred-set_username}{\code{AuthCred$set_username()}}
\item \href{#method-AuthCred-get_username}{\code{AuthCred$get_username()}}
\item \href{#method-AuthCred-set_password}{\code{AuthCred$set_password()}}
\item \href{#method-AuthCred-clear_password}{\code{AuthCred$clear_password()}}
\item \href{#method-AuthCred-get_password}{\code{AuthCred$get_password()}}
\item \href{#method-AuthCred-has_cred}{\code{AuthCred$has_cred()}}
\item \href{#method-AuthCred-clone}{\code{AuthCred$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthCred-new"></a>}}
\if{latex}{\out{\hypertarget{method-AuthCred-new}{}}}
\subsection{Method \code{new()}}{
Create a new AuthCred
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthCred$new(username = NA_character_, password = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{username}}{The KHIS username.}

\item{\code{password}}{The KHIS password.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthCred-set_username"></a>}}
\if{latex}{\out{\hypertarget{method-AuthCred-set_username}{}}}
\subsection{Method \code{set_username()}}{
Set the KHIS username
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthCred$set_username(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The KHIS username}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthCred-get_username"></a>}}
\if{latex}{\out{\hypertarget{method-AuthCred-get_username}{}}}
\subsection{Method \code{get_username()}}{
Get username
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthCred$get_username()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthCred-set_password"></a>}}
\if{latex}{\out{\hypertarget{method-AuthCred-set_password}{}}}
\subsection{Method \code{set_password()}}{
Set the KHIS password
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthCred$set_password(value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{value}}{The KHIS password}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthCred-clear_password"></a>}}
\if{latex}{\out{\hypertarget{method-AuthCred-clear_password}{}}}
\subsection{Method \code{clear_password()}}{
Clear password
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthCred$clear_password()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthCred-get_password"></a>}}
\if{latex}{\out{\hypertarget{method-AuthCred-get_password}{}}}
\subsection{Method \code{get_password()}}{
Get password
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthCred$get_password()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthCred-has_cred"></a>}}
\if{latex}{\out{\hypertarget{method-AuthCred-has_cred}{}}}
\subsection{Method \code{has_cred()}}{
Report if we have credentials
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthCred$has_cred()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-AuthCred-clone"></a>}}
\if{latex}{\out{\hypertarget{method-AuthCred-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{AuthCred$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
