% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerscreening-package.R
\docType{package}
\name{cancerscreening-package}
\alias{cancerscreening}
\alias{cancerscreening-package}
\title{cancerscreening: Streamline Access to Cancer Screening Data}
\description{
cancerscreening provide a easy way to download cancer screening data from the
Kenya Health Information System (KHIS) using R.

Most function begin with the prefix \code{get_} followed by the screening area
\code{cervical} or \code{breast}. The Goal is to allow the download of data associated
with the data of interest, e.g. \code{get_cervical_screened}, \code{get_cervical_positive},
or \code{get_cervical_treated}.

cancerscreening is "pipe-friendly" and, in fact re-exports \verb{\%>\%} but does not
require its use.

Please see the cancerscreening website for full documentation:
\itemize{
\item \url{https://cancerscreening.damurka.com/index.html}
}

In addition to function-specific help, there are several articles which are
indexed here:
\itemize{
\item \href{https://cancerscreening.damurka.com/articles/index.html}{Article index}
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://cancerscreening.damurka.com}
  \item \url{https://github.com/damurka/cancerscreening}
  \item Report bugs at \url{https://github.com/damurka/cancerscreening/issues}
}

}
\author{
\strong{Maintainer}: David Kariuki \email{hello@damurka.com} (\href{https://orcid.org/0009-0003-6159-1107}{ORCID}) [copyright holder]

}
\keyword{internal}
