% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{cbe_target_population}
\alias{cbe_target_population}
\title{Get the target population for clinical breast examination in Kenya}
\usage{
cbe_target_population(year, includeSubCounty = FALSE)
}
\arguments{
\item{year}{Year for which to estimate population}

\item{includeSubCounty}{Whether to include sub county data (default: false)}
}
\value{
A tibble containing the target population for clinical breast examination (CBE)
\describe{
   \item{county}{name of the county}
   \item{target}{number to be screened}
}
}
\description{
\code{cbe_target_population} function subsets the total population produced by
\link{total_population} function and creates a target population eligible for
clinical breast examination: females aged between 25 years and 74 years
}
\details{
Implements the \link{breast_target_population} with the min_age set at 25 and
max_age at 75 as guided by the Kenya National Cancer Screening Guideline 2018
}
\author{
David Kariuki
}
