% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{cervical_target_population}
\alias{cervical_target_population}
\title{Get the target population for cervical cancer screening in Kenya}
\usage{
cervical_target_population(year, includeSubCounty = FALSE)
}
\arguments{
\item{year}{Year for which to estimate population}

\item{includeSubCounty}{Whether to include sub county data (default: false)}
}
\value{
A tibble containing the target population for cervical cancer screening
\describe{
   \item{county}{name of the county}
   \item{target}{number to be screened}
}
}
\description{
\code{cervical_target_population} function subsets the Kenyan population and
creates a target population for cervical cancer screening: females aged
between 25 years and 50 years
}
\details{
The target population for cervical cancer screening is determined by the Kenya
National Cancer Screening guidelines 2018 and follows the guidance outlined in
the WHO publication 'Planning and implementing cervical cancer prevention programs:
A manual for managers.' The annual screening targets for counties and the national
level are calculated based on population growth. The formula can be found in
\link{population_growth}
}
\author{
David Kariuki
}
