% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cervical_data.R
\name{.get_cervical_data}
\alias{.get_cervical_data}
\title{Retrieves cervical cancer screening and treatment data}
\usage{
.get_cervical_data(
  element_ids,
  start_date,
  end_date = NULL,
  level = c("kenya", "county", "subcounty", "ward", "facility"),
  organisations = NULL,
  categories = NULL,
  elements = NULL,
  khis_session = dynGet("khis_default_session", inherits = TRUE),
  retry = 1,
  verbosity = 0
)
}
\arguments{
\item{element_ids}{A vector of data element IDs to retrieve data for.}

\item{start_date}{The start date for the data retrieval in the format 'YYYY-MM-dd'.}

\item{end_date}{The end date for the data retrieval in the format 'YYYY-MM-dd'. Defaults to the current date if not specified.}

\item{level}{The desired level of the organization unit hierarchy to retrieve data for: "kenya", "county", "subcounty", "ward", or "facility".}

\item{organisations}{A vector of specific organisation unit IDs to retrieve data for. If NULL, data for all organisation units at the specified level will be retrieved.}

\item{categories}{A tibble of category options, typically obtained from \code{get_categories()}. If NULL, categories will be downloaded automatically.}

\item{elements}{A tibble of data elements, typically obtained from \code{get_data_elements()}. If NULL, elements will be downloaded automatically.}

\item{khis_session}{The KHIS session object to use (defaults to "khis_default_session"). See \code{?login_to_khis()} for details.}

\item{retry}{Number of times to retry the API call in case of failure (defaults to 1).}

\item{verbosity}{Level of information to print during the API call:
\itemize{
\item 0: No output
\item 1: Show headers
\item 2: Show headers and bodies
\item 3: Show headers, bodies, and curl status message
}}
}
\value{
A tibble containing cervical screening and treatment data with the following columns:
\describe{
\item{facility}{Name of the health facility.}
\item{ward}{Name of the ward.}
\item{subcounty}{Name of the subcounty.}
\item{county}{Name of the county.}
\item{period}{Date of the report.}
\item{month}{Month of the report.}
\item{year}{Year of the report.}
\item{fiscal_year}{Financial year of the report (July-June cycle).}
\item{category}{Age group category (<25, 25-49, 50+).}
\item{category2}{Additional category option (if applicable).}
\item{element}{Name of the data element.}
\item{source}{Source document of the report (MOH 711 or MOH 745).}
\item{value}{Numeric value for the number of cases reported.}
}
}
\description{
\code{.get_cervical_data()} retrieves cervical cancer screening and treatment data for a specified period from the KHIS API server.
}
\seealso{
get_analytics(), get_categories(), get_data_elements(), login_to_khis()
}
\keyword{internal}
