% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{facility_ownership}
\alias{facility_ownership}
\title{Facility Ownership Metadata}
\format{
A tibble with 2 columns and 4 rows
\itemize{
\item ownership_id - unique identifier for the data element.
\item facility_ownership - name of the data element.
}
}
\source{
Kenya Health Information System \url{https://hiskenya.org}
}
\usage{
facility_ownership
}
\description{
Provides the facility ownership metadata.
}
\details{
The data will contain the following info about the facility ownership organisation unit group
\itemize{
\item JlW9OiK1eR4 - Facility Ownership
\item eT1vvFVhLHc - Faith Based Organisation
\item AaAF5EmS1fk - Ministry of Health
\item aRxa6o8GqZN - Private
}
}
\keyword{internal}
