% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{facility_type}
\alias{facility_type}
\title{Facility Type Metadata}
\format{
A tibble with 2 columns and 7 rows
\itemize{
\item type_id        - unique identifier for the data element
\item facility_type  - name of the data element
}
}
\source{
Kenya Health Information System \url{https://hiskenya.org}
}
\usage{
facility_type
}
\description{
Provides the facility type metadata.
}
\details{
The data will contain the following info about the facility type organisation unit group
\itemize{
\item FSoqQFDES0U - Facility Type
\item mVrepdLAqSD - Revised Dispensary (2018)
\item rhKJPLo27x7 - Revised Health Centre (2018)
\item CGDNIWGHRNr - Revised Hospital (2018)
\item xKv918PvmHN - Revised Medical Center (2018)
\item lTrpyOiOcM6 - Revised Medical Clinic (2018)
\item lQXBYwznvXA - Revised Nursing Home (2018)
\item YQK9pleIoeB - Revised Stand alone(2018)
}
}
\keyword{internal}
