% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_population.R
\name{get_breast_cbe_target_population}
\alias{get_breast_cbe_target_population}
\title{Get the Clinical Breast Examination Target Population}
\usage{
get_breast_cbe_target_population(
  year,
  level = c("kenya", "county", "subcounty")
)
}
\arguments{
\item{year}{Year for which to estimate population}

\item{level}{The desired level of the organization unit hierarchy to retrieve
data for: \code{"kenya"} (default) , \code{"county"} or \code{"subcounty"}.}
}
\value{
A tibble containing the target population for clinical breast
examination (CBE)
\itemize{
\item county     - name of the county. Optional if the level is county or subcounty
\item subcounty  - name of the county. Optional if the level if subcounty
\item target     - number to be screened
}
}
\description{
\code{get_breast_cbe_target_population()} subsets the Kenyan population creates a
target population eligible for clinical breast examination: females aged
between 25 years and 74 years
}
\details{
Implements age group with the min_age set at 25 and max_age at 75 as
guided by the Kenya National Cancer Screening Guideline 2018
}
\examples{
# Get the country projection of women to perform CBE for the year 2024
target_population <- get_breast_cbe_target_population(2024)
target_population

# Get the projection for CBE for 2022 by county
target_population <- get_breast_cbe_target_population(2022, level = 'county')
target_population
}
