% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_population.R
\name{get_breast_mammogram_target_population}
\alias{get_breast_mammogram_target_population}
\title{Get Mammography Target Population}
\usage{
get_breast_mammogram_target_population(
  year,
  level = c("kenya", "county", "subcounty")
)
}
\arguments{
\item{year}{Year for which to estimate population}

\item{level}{The desired level of the organization unit hierarchy to retrieve
data for: \code{"kenya"} (default) , \code{"county"} or \code{"subcounty"}.}
}
\value{
A tibble containing the target population eligible for breast cancer
\itemize{
\item county     - name of the county. Optional if the level is county or subcounty
\item subcounty  - name of the county. Optional if the level if subcounty
\item target     - number to be screened
}
}
\description{
\code{get_breast_mammogram_target_population()} subsets the Kenyan population and
creates a target population eligible for breast cancer screening through
mammography: females aged between 40 years to 74 years
}
\details{
Implements the age group population with the min_age set at 40 and
max_age at 75 as guided by the Kenya National Cancer Screening Guideline 2018
}
\examples{
# Get the country projection of women to perform mammogram for the year 2024
target_population <- get_breast_mammogram_target_population(2024)
target_population

# Get the projection for mammograms for 2022 by county
target_population <- get_breast_mammogram_target_population(2022, level = 'county')
target_population
}
