% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_categories.R
\name{get_categories}
\alias{get_categories}
\title{Get a list of category options}
\usage{
get_categories(
  khis_session = dynGet("khis_default_session", inherits = TRUE),
  retry = 1,
  verbosity = 0
)
}
\arguments{
\item{khis_session}{The KHIS session object to use (defaults to "khis_default_session"). See \code{?login_to_khis()} for details.}

\item{retry}{Number of times to retry the API call in case of failure (defaults to 1).}

\item{verbosity}{Level of information to print during the API call:
\itemize{
\item 0: No output
\item 1: Show headers
\item 2: Show headers and bodies
\item 3: Show headers, bodies, and curl status message
}}
}
\value{
A tibble containing the following columns:
\describe{
\item{category_id}{The unique identifier for the category option.}
\item{category}{The name of the category option.}
}
}
\description{
\code{get_categories()} retrieves a list of category options from the KHIS API
server, including their IDs and names.
}
\seealso{
\code{\link[=login_to_khis]{login_to_khis()}}
}
