% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_category_options_metadata.R
\name{get_category_options_metadata}
\alias{get_category_options_metadata}
\title{Get Category Options Metadata}
\usage{
get_category_options_metadata(...)
}
\arguments{
\item{...}{
  Arguments passed on to \code{\link[khisr:get_metadata]{khisr::get_metadata}}
  \describe{
    \item{\code{fields}}{The specific columns to be returned in the tibble.}
    \item{\code{retry}}{Number of times to retry the API call in case of failure
(defaults to 2).}
    \item{\code{verbosity}}{Level of HTTP information to print during the call:
\itemize{
\item 0: No output
\item 1: Show headers
\item 2: Show headers and bodies
\item 3: Show headers, bodies, and CURL status message.
}}
    \item{\code{timeout}}{Maximum number of seconds to wait for the API response.}
  }}
}
\value{
A tibble containing the following columns:
\itemize{
\item category_id  - The unique identifier for the category option.
\item category     - The name of the category option.
}
}
\description{
\code{get_category_options_metadata()} fetches category options from the KHIS API
}
\examples{
\dontshow{if (khis_has_cred()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Fetch the category option metadata
categories <- get_category_options_metadata()
categories
\dontshow{\}) # examplesIf}
}
