% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breast_screening.R
\name{get_cbe_conducted}
\alias{get_cbe_conducted}
\title{Retrieves data for clinical breast examination (CBE) conducted}
\usage{
get_cbe_conducted(
  start_date,
  end_date = NULL,
  categories = NULL,
  facilities = NULL,
  elements = NULL
)
}
\arguments{
\item{start_date}{The start date for the data retrieval in the format 'YYYY-MM-dd'}

\item{end_date}{The end date for the data retrieval in the format 'YYYY-MM-dd'. The default is to get the current date}

\item{categories}{The list of categories. The default action is download using \link{get_categories}}

\item{facilities}{The list of facilities. The default action is download using \link{get_facilities}}

\item{elements}{The list of data elements. The default action is download using \link{get_data_elements}}
}
\value{
A tibble containing a list of clinical breast examinations conducted
\describe{
   \item{facility_id}{organisation identifier that uniquely identifies the health facility}
   \item{facility}{name of the health facility}
   \item{facility_type}{The type of the facility: Dispensary, health centre, medical clinic etc.}
   \item{facility_ownership}{facility ownership: faith based, private or governement}
   \item{keph_level}{facility level}
   \item{county}{name of the county}
   \item{subcounty}{name of the subcounty}
   \item{month}{name of the month reported}
   \item{year}{the year of the report}
   \item{fiscal_year}{the financial year of the report. uses July - June Cycle}
   \item{category}{name of the category option}
   \item{element}{name of the data element}
   \item{source}{source document of the report}
   \item{element}{name of the data element}
   \item{period}{the month of the report}
   \item{value}{The numeric value for the number of cases reported}
}
}
\description{
\code{get_cbe_conducted} function retrieves data for CBE conducted within a specified
period using the \link{get_breast_data} function.
}
