% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cbe_target_population.R
\name{get_cbe_target_population}
\alias{get_cbe_target_population}
\title{Get the target population for clinical breast examination in Kenya}
\usage{
get_cbe_target_population(year, level = c("kenya", "county", "subcounty"))
}
\arguments{
\item{year}{Year for which to estimate population}

\item{level}{The desired level of the organization unit hierarchy to retrieve data for: "kenya", "county" or "subcounty".}
}
\value{
A tibble containing the target population for clinical breast
examination (CBE)
\describe{
\item{kenya}{name of the country}
\item{county}{name of the county}
\item{subcounty}{name of the county}
\item{target}{number to be screened}
}
}
\description{
\code{get_cbe_target_population} function subsets the Kenyan population creates a
target population eligible for clinical breast examination: females aged
between 25 years and 74 years
}
\details{
Implements age group with the min_age set at 25 and max_age at 75 as guided by
the Kenya National Cancer Screening Guideline 2018
}
