% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cervical_screening.R
\name{get_cervical_data}
\alias{get_cervical_data}
\title{Retrieves cervical cancer screen and treatment data}
\usage{
get_cervical_data(
  element_ids,
  start_date,
  end_date = NULL,
  categories = NULL,
  facilities = NULL,
  elements = NULL,
  d2_session = dynGet("d2_default_session", inherits = TRUE)
)
}
\arguments{
\item{element_ids}{A vector with the list of data element  to retrieve data}

\item{start_date}{The start date for the data retrieval in the format 'YYYY-MM-dd'}

\item{end_date}{The end date for the data retrieval in the format 'YYYY-MM-dd'. The default is to get the current date}

\item{categories}{The list of categories. The default action is download using \link{get_categories}}

\item{facilities}{The list of facilities. The default action is download using \link{get_facilities}}

\item{elements}{The list of data elements. The default action is download using \link{get_data_elements}}

\item{d2_session}{the khisSession object, default is "d2_default_session",
it will be made upon logining in to KHIS with \link{loginToKHIS}}
}
\value{
A tibble containing cervical screening and treatment data
\describe{
   \item{facility_id}{organisation identifier that uniquely identifies the health facility}
   \item{facility}{name of the health facility}
   \item{facility_type}{The type of the facility: Dispensary, health centre, medical clinic etc.}
   \item{facility_ownership}{facility ownership: faith based, private or governement}
   \item{keph_level}{facility level}
   \item{county}{name of the county}
   \item{subcounty}{name of the subcounty}
   \item{month}{name of the month reported}
   \item{year}{the year of the report}
   \item{fiscal_year}{the financial year of the report. uses July - June Cycle}
   \item{category}{name of the category option}
   \item{element}{name of the data element}
   \item{source}{source document of the report}
   \item{element}{name of the data element}
   \item{period}{the month of the report}
   \item{value}{The numeric value for the number of cases reported}
}
}
\description{
\code{get_cervical_data} function retrieves cervical cancer screening and treatment
 data for a specified period using the \link{get_analytics} function.
}
