% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cervical_positive.R
\name{get_cervical_positive}
\alias{get_cervical_positive}
\title{Retrieves Cervical Cancer Screening Data with Positive Results}
\usage{
get_cervical_positive(
  start_date,
  end_date = NULL,
  level = c("kenya", "county", "subcounty", "ward", "facility"),
  organisations = NULL,
  categories = NULL,
  elements = NULL,
  ...
)
}
\arguments{
\item{start_date}{The start date to retrieve data. It is required and in the format 'YYYY-MM-dd'.}

\item{end_date}{The ending date for data retrieval (default is the current date).}

\item{level}{The desired data granularity: \code{"kenya"} (the default), \code{"county"}, \code{"subcounty"}, \code{"ward"}, or \code{"facility"}.}

\item{organisations}{A list of organization units in the data. If NULL, downloaded using \code{\link[=get_organisation_units_metadata]{get_organisation_units_metadata()}}.}

\item{categories}{A list of categories to include. If NULL, downloaded using \code{\link[=get_category_options_metadata]{get_category_options_metadata()}}.}

\item{elements}{A list of data elements to include. If NULL, downloaded using \code{\link[=get_data_elements_metadata]{get_data_elements_metadata()}}.}

\item{...}{Other options that can be passed onto \link{.api_get}.}
}
\value{
A tibble containing cervical cancer screening data with positive results
with the following columns:
}
\description{
\code{get_cervical_positive()} retrieves cervical cancer screening data with positive results
for a specified period from the KHIS API server using \code{\link[=.get_cervical_data]{.get_cervical_data()}}.
}
\examples{
\dontshow{if (khis_has_cred()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Download data from February 2023 to current date
data <- get_cervical_positive(start_date = '2023-02-01')
data
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=.get_cervical_data]{.get_cervical_data()}} for retrieving and formatting cervical cancer screening data
}
