% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_population.R
\name{get_cervical_target_population}
\alias{get_cervical_target_population}
\title{Get Cervical Cancer Screening Target Population}
\usage{
get_cervical_target_population(year, level = c("kenya", "county", "subcounty"))
}
\arguments{
\item{year}{Year for which to estimate population}

\item{level}{The desired level of the organization unit hierarchy to retrieve
data for: \code{"kenya"} (default) , \code{"county"} or \code{"subcounty"}.}
}
\value{
A tibble containing the target population for cervical cancer
screening
\itemize{
\item county     - name of the county. Optional if the level is county or subcounty
\item subcounty  - name of the county. Optional if the level if subcounty
\item target     - number to be screened
}
}
\description{
\code{get_cervical_target_population()} subsets the Kenyan population and creates
a target population for cervical cancer screening: females aged between 25
years and 50 years
}
\details{
The target population for cervical cancer screening is determined by
the Kenya National Cancer Screening guidelines 2018 and follows the guidance
outlined in the WHO publication 'Planning and implementing cervical cancer
prevention programs: A manual for managers.' The annual screening targets for
counties and the national level are calculated based on population growth
}
\examples{
# Get the country projection for cervical cancer screening for the year 2024
target_population <- get_cervical_target_population(2024)
target_population

# Get the projection for cervical cancer screening for 2022 by county
target_population <- get_cervical_target_population(2022, level = 'county')
target_population
}
