% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_facilities}
\alias{get_facilities}
\title{Get list of health facilities in Kenya}
\usage{
get_facilities()
}
\value{
A tibble containing a list of health facilities in Kenya
\describe{
   \item{facility_id}{organisation identifier that uniquely identifies the health facility}
   \item{facility}{name of the health facility}
   \item{county}{name of the county}
   \item{subcounty}{name of the subcounty}
   \item{ward}{name of the ward}
}
}
\description{
\code{get_facilities} function retrieves a list of health facilities or organisation
units using the \link{api_get} function. The organisation unit only filters for
health facilities only
}
