% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_facility_ownerships}
\alias{get_facility_ownerships}
\title{Get the facility ownership metadata}
\usage{
get_facility_ownerships()
}
\value{
A tibble containing a list of facility ownership metadata
\describe{
   \item{ownership_id}{unique identifier for the data element}
   \item{facility_ownership}{name of the data element}
}
}
\description{
\code{get_facility_ownerships} function retrieves the facility ownership metadata
}
