% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_facility_types}
\alias{get_facility_types}
\title{Get the facility type metadata}
\usage{
get_facility_types()
}
\value{
A tibble containing a list of facility type metadata
\describe{
   \item{type_id}{unique identifier for the data element}
   \item{facility_type}{name of the data element}
}
}
\description{
\code{get_facility_types} function retrieves the facility type metadata
}
