% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_facility_metaadata.R
\name{get_keph_levels}
\alias{get_keph_levels}
\title{Get the facility Kenya essential package for Health (KEPH) level metadata}
\usage{
get_keph_levels()
}
\value{
A tibble containing a list of facility KEPH level metadata
\describe{
\item{keph_id}{unique identifier for the data element}
\item{keph_level}{name of the data element}
}
}
\description{
\code{get_keph_levels} function retrieves the facility KEPH level metadata
}
