% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_mammogram_target_population.R
\name{get_mammogram_target_population}
\alias{get_mammogram_target_population}
\title{Get the target population for breast cancer screening through mammography}
\usage{
get_mammogram_target_population(
  year,
  level = c("kenya", "county", "subcounty")
)
}
\arguments{
\item{year}{Year for which to estimate population}

\item{level}{The desired level of the organization unit hierarchy to retrieve data for: "kenya", "county" or "subcounty".}
}
\value{
A tibble containing the target population eligible for breast cancer
\describe{
\item{kenya}{name of the country}
\item{county}{name of the county}
\item{subcounty}{name of the county}
\item{target}{number to be screened}
}
}
\description{
\code{get_mammogram_target_population} subsets the Kenyan population and creates a
target population eligible for breast cancer screening through mammography:
females aged between 40 years to 74 years
}
\details{
Implements the age group population with the min_age set at 40 and max_age at
75 as guided by the Kenya National Cancer Screening Guideline 2018
}
