% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_organisation_units.R
\name{get_organisation_units}
\alias{get_organisation_units}
\title{Get a list of organisation units in KHIS}
\usage{
get_organisation_units(
  khis_session = dynGet("khis_default_session", inherits = TRUE),
  retry = 1,
  verbosity = 0
)
}
\arguments{
\item{khis_session}{The KHIS session object to use (defaults to "khis_default_session").
See \code{?login_to_khis()} for details.}

\item{retry}{Number of times to retry the API call in case of failure (defaults to 1).}

\item{verbosity}{Level of information to print during the API call:
\itemize{
\item 0: No output
\item 1: Show headers
\item 2: Show headers and bodies
\item 3: Show headers, bodies, and curl status message
}}
}
\value{
A tibble containing the following columns:
\describe{
\item{facility_id}{Organisation identifier that uniquely identifies the health facility.}
\item{ward_id}{Organisation identifier that uniquely identifies the ward.}
\item{subcounty_id}{Organisation identifier that uniquely identifies the subcounty.}
\item{county_id}{Organisation identifier that uniquely identifies the county.}
\item{facility}{Name of the health facility.}
\item{county}{Name of the county.}
\item{subcounty}{Name of the subcounty.}
\item{ward}{Name of the ward.}
}
}
\description{
\code{\link[=get_organisation_units]{get_organisation_units()}} retrieves a list of organisation units from the
KHIS API server, including their names, counties, subcounties, and wards.
}
\seealso{
\code{\link[=login_to_khis]{login_to_khis()}}
}
