% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{getCredentialsFromKeyring}
\alias{getCredentialsFromKeyring}
\title{getCredentialsFromKeyring(ring, service, username)}
\usage{
getCredentialsFromKeyring(username)
}
\arguments{
\item{username}{KHIS username}
}
\value{
a list containing entries called password and username
}
\description{
retrieves username, service, and password from keyring
}
