% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AuthCred-class.R
\name{init_AuthCred}
\alias{init_AuthCred}
\title{Create an AuthCred}
\usage{
init_AuthCred(username = NA_character_, password = NULL)
}
\arguments{
\item{username}{The KHIS username to be used in API calls}

\item{password}{The KHIS password to be used in API calls. Typically
populated indirectly if stored using the \link{keyring}}
}
\value{
An object of class \link{AuthCred}
}
\description{
Constructor function for objects of class \link{AuthCred}.
}
\examples{
init_AuthCred(
  username = 'khis_username'
)

}
\keyword{internal}
