% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{keph_level}
\alias{keph_level}
\title{Kenya essential package for Health (KEPH) Level Metadata}
\format{
A tibble with 2 columns and 6 rows
\itemize{
\item keph_id    - unique identifier for the data element
\item keph_level - name of the data element
}
}
\source{
Kenya Health Information System \url{https://hiskenya.org}
}
\usage{
keph_level
}
\description{
Provided the facility KEPH level metadata.
}
\details{
The data will contain the following info about the KEPH organisation unit group
\itemize{
\item sytI3XYbxwE - KEPH Level
\item axUnguN4QDh - KEPH Level 1
\item tvMxZ8aCVou - KEPH Level 2
\item wwiu1jyZOXO - KEPH Level 3
\item hBZ5DRto7iF - KEPH Level 4
\item d5QX71PY5t0 - KEPH Level 5
\item FpY8vg4gh46 - KEPH Level 6
}
}
\keyword{internal}
