% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khis_cred.R
\name{khis_cred}
\alias{khis_cred}
\title{Sets the KHIS Credentials}
\usage{
khis_cred(config_path = NULL, username = NULL, password = NULL)
}
\arguments{
\item{config_path}{An optional parameters that contains the path to configuration file with username and password}

\item{username}{The KHIS username. Can be optional if \code{config_path} is already provided}

\item{password}{The KHIS password. Can be optional if \code{config_path} is already provided or already stored using the \link{keyring} package}
}
\description{
Sets the KHIS Credentials
}
\examples{
\dontshow{if (rlang::is_interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Load from a configuration file
khis_cred(config_path = 'cred.json')

# Load username and password
khis_cred(username = 'damurka', password = 'PASSWORD')

# Load if the password has been used before
khis_cred(username = 'damurka')
\dontshow{\}) # examplesIf}
}
\concept{cred functions}
