% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khis_cred.R
\name{khis_cred_internal}
\alias{khis_cred_internal}
\title{Internal Credentials}
\usage{
khis_cred_internal(account = c("docs", "testing"))
}
\arguments{
\item{account}{The environment to provide credentials. \code{"docs"} or \code{"testing"}}
}
\description{
Internal function used to provide credentials for the testing and documentation
environment
}
\keyword{internal}
