% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khis_cred.R
\name{khis_has_cred}
\alias{khis_has_cred}
\title{Are There Credentials on Hand?}
\usage{
khis_has_cred()
}
\description{
Are There Credentials on Hand?
}
\examples{
khis_has_cred()
}
\concept{low-level API functions}
