% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khis_cred.R
\name{khis_username}
\alias{khis_username}
\title{Produces the Configured Username}
\usage{
khis_username()
}
\value{
the username of the user credentials
}
\description{
Produces the Configured Username
}
\examples{
rlang::is_interactive()
\dontrun{
  # Set the credentials
  khis_cred(username = 'user', password = 'password')

  # View the username
  khis_username()

  # Clear credentials
  khis_clear_cred()
}
}
\seealso{
Other low-level API functions: 
\code{\link{khis_has_cred}()}
}
\concept{low-level API functions}
