% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{khisSession}
\alias{khisSession}
\title{khisSession}
\description{
khisSession

khisSession
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{config_path}}{Path to a JSON configuration
file.}

\item{\code{username}}{The KHIS username.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-khisSession-new}{\code{khisSession$new()}}
\item \href{#method-khisSession-clone}{\code{khisSession$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-khisSession-new"></a>}}
\if{latex}{\out{\hypertarget{method-khisSession-new}{}}}
\subsection{Method \code{new()}}{
Create a new KHISLogin object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{khisSession$new(config_path = NA_character_, username)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{config_path}}{Configuration file path}

\item{\code{username}}{URL to the server.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-khisSession-clone"></a>}}
\if{latex}{\out{\hypertarget{method-khisSession-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{khisSession$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
