% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/login.R
\name{loadConfigFile}
\alias{loadConfigFile}
\title{LoadConfig(config_path)}
\usage{
loadConfigFile(config_path = NA)
}
\arguments{
\item{config_path}{Path to the KHIS credentials file}
}
\value{
A parsed list of the configuration file.
}
\description{
Loads a JSON configuration file to access a KHIS instance
}
