% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{mammogram_target_population}
\alias{mammogram_target_population}
\title{Get the target population for breast cancer screening through mammography}
\usage{
mammogram_target_population(year, includeSubCounty = FALSE)
}
\arguments{
\item{year}{Year for which to estimate population}

\item{includeSubCounty}{Whether to include sub county data (default: false)}
}
\value{
A tibble containing the target population eligible for breast cancer
screening through mammography
\describe{
\item{county}{name of the county}
\item{target}{number to be screened}
}
}
\description{
\code{mammo_target_population} subsets the Kenyan population and creates a
target population eligible for breast cancer screening through mammography:
females aged between 40 years to 74 years
}
\details{
Implements the \link{breast_target_population} with the min_age set at 40 and
max_age at 75 as guided by the Kenya National Cancer Screening Guideline 2018
}
