% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{population_growth}
\alias{population_growth}
\title{Projects population growth}
\usage{
population_growth(P, year, rate = 0.022)
}
\arguments{
\item{P}{Initial population size}

\item{year}{Year for which to estimate population}

\item{rate}{Annual growth rate}
}
\value{
Projected population size at the specified year
}
\description{
\code{population_growth} function projects population growth based on the formula
\code{Nt = P * e^(r * t)}. The annual population growth rate, r, is set at 2.2% and
the years,t, calculated with the reference to the year 2020 year after the last
population census
}
