% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexport-khis_cred.R, R/reexport-magrittr.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{khis_cred}
\alias{khis_has_cred}
\alias{khis_cred_clear}
\alias{khis_username}
\alias{khis_base_url}
\alias{\%>\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{khisr}{\code{\link[khisr]{khis_base_url}}, \code{\link[khisr]{khis_cred}}, \code{\link[khisr]{khis_cred_clear}}, \code{\link[khisr]{khis_has_cred}}, \code{\link[khisr]{khis_username}}}

  \item{magrittr}{\code{\link[magrittr:pipe]{\%>\%}}}
}}

