% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khis_cred.R
\name{req_auth_khis_basic}
\alias{req_auth_khis_basic}
\title{Authenticate Request with HTTP Basic Authentication}
\usage{
req_auth_khis_basic(req)
}
\arguments{
\item{req}{A request}
}
\value{
A modified HTTP request with authorization header
}
\description{
This sets the Authorization header for basic authentication using the username and password provided.
}
\examples{
\dontshow{if (khis_has_cred()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
req <- request("http://example.com") |> req_auth_khis_basic("damurka", "PASSWORD")
req
\dontshow{\}) # examplesIf}
}
\seealso{
\link{httr2}
}
