% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{total_population}
\alias{total_population}
\title{Get Kenya 2019 census population}
\usage{
total_population()
}
\value{
A list containing data frame for each population:
\describe{
   \item{county}{name of the county}
   \item{subcounty}{name of the subcounty}
   \item{age}{age of the population in years}
   \item{sex}{sex (male or female)}
   \item{population}{number of people}
}
}
\description{
\code{total_population} retrieves the Kenyan population data by county, subcounty,
sex, and age based on the 2019 population census.
}
