% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{api_get}
\alias{api_get}
\title{Make an API call to the Kenya Health Information System (KHIS) server}
\usage{
api_get(url_path, ..., retry = NULL)
}
\arguments{
\item{url_path}{The path to make an api call}

\item{...}{Name-value pairs that provide query parameters. Each value must be
either a length-1 atomic vector (which is automatically escaped) or NULL (which
is silently dropped).}

\item{retry}{The maximum number of attempts retries. If not supplied (the default),it will not retry.}
}
\value{
Parsed JSON
}
\description{
\code{api_get} function makes an API call to the KHIS server
}
