% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/populations.R
\name{breast_target_population}
\alias{breast_target_population}
\title{Get the target population for breast cancer screening}
\usage{
breast_target_population(year, min_age, max_age = 75, includeSubCounty = FALSE)
}
\arguments{
\item{year}{Year for which to estimate population}

\item{min_age}{The minimum age to be in the population}

\item{max_age}{The maximum age to be in the population (not included)}

\item{includeSubCounty}{Whether to include sub county data (default: false)}
}
\value{
A tibble containing the target population eligible for breast cancer screening
\describe{
   \item{county}{name of the county}
   \item{target}{number to be screened}
}
}
\description{
\code{breast_target_population} subsets the total population produced by
\link{total_population} function and creates a target population eligible for
breast cancer screening
}
\details{
Breast cancer screening target population based on Kenya National Cancer Screening
guidelines 2018. Annual targets increase from 5% in 2021 to 42% in 2030 Population
growth is calculated using: \link{population_growth}. The annual population growth
rate, is set at 2.2% and the years calculated with the reference to the year 2020
year after the last population census
}
\author{
David Kariuki
}
