% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancerscreening-package.R
\name{cancerscreening-configuration}
\alias{cancerscreening-configuration}
\title{cancerscreening configuration}
\description{
Some aspects of cancerscreening behaviour can be controlled via an option.
}
\section{Auth}{


Read about cancerscreening's main auth function, \code{\link[=khis_cred]{khis_cred()}}. It is stores
the KHIS credentials for authorization during API calls.
}

