% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_categories}
\alias{get_categories}
\title{Get list of category options metadata}
\usage{
get_categories()
}
\value{
A tibble containing a list of category options
\describe{
   \item{category_id}{unique identifier for the category option}
   \item{category}{name of the category option}
}
}
\description{
\code{get_categories} function retrieves a list of category options using the
\link{api_get} function.
}
