% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_categories}
\alias{get_categories}
\title{Get list of category options metadata}
\usage{
get_categories(d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2_session}{the khisSession object, default is "d2_default_session"}
}
\value{
A tibble containing a list of category options
\describe{
   \item{category_id}{unique identifier for the category option}
   \item{category}{name of the category option}
}
}
\description{
\code{get_categories} function retrieves a list of category options using the
\link{api_get} function.
}
