% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_category_options_metadata.R
\name{get_category_options_metadata}
\alias{get_category_options_metadata}
\title{Get Category Options Metadata}
\usage{
get_category_options_metadata(...)
}
\arguments{
\item{...}{Options that can be passed onto KHIS API.}
}
\value{
A tibble containing the following columns:
\itemize{
\item category_id  - The unique identifier for the category option.
\item category     - The name of the category option.
}
}
\description{
\code{get_category_options_metadata()} fetches category options from the KHIS API
}
\examples{
\dontshow{if (khis_has_cred()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Fetch the category option metadata
categories <- get_category_options_metadata()
categories
\dontshow{\}) # examplesIf}
}
