% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cervical_target_population.R
\name{get_cervical_target_population}
\alias{get_cervical_target_population}
\title{Get the target population for cervical cancer screening in Kenya}
\usage{
get_cervical_target_population(year, level = c("kenya", "county", "subcounty"))
}
\arguments{
\item{year}{Year for which to estimate population}

\item{level}{The desired level of the organization unit hierarchy to retrieve data for: "kenya", "county" or "subcounty".}
}
\value{
A tibble containing the target population for cervical cancer screening
\describe{
\item{kenya}{name of the country}
\item{county}{name of the county}
\item{subcounty}{name of the county}
\item{target}{number to be screened}
}
}
\description{
\code{get_cervical_target_population} function subsets the Kenyan population and
creates a target population for cervical cancer screening: females aged
between 25 years and 50 years
}
\details{
The target population for cervical cancer screening is determined by the Kenya
National Cancer Screening guidelines 2018 and follows the guidance outlined in
the WHO publication 'Planning and implementing cervical cancer prevention programs:
A manual for managers.' The annual screening targets for counties and the national
level are calculated based on population growth
}
