% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parser.R
\name{get_data_elements}
\alias{get_data_elements}
\title{Get list of data elements metadata}
\usage{
get_data_elements(d2_session = dynGet("d2_default_session", inherits = TRUE))
}
\arguments{
\item{d2_session}{the khisSession object, default is "d2_default_session",}
}
\value{
A tibble containing a list of data elements
\describe{
   \item{element_id}{unique identifier for the data element}
   \item{element}{name of the data element}
}
}
\description{
\code{get_data_elements} function retrieves a list of data elements using the
\link{api_get} function.
}
