% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data_elements.R
\name{get_data_elements}
\alias{get_data_elements}
\title{Get a list of data elements}
\usage{
get_data_elements(
  element_ids = NULL,
  khis_session = dynGet("khis_default_session", inherits = TRUE),
  retry = 1,
  verbosity = 0
)
}
\arguments{
\item{element_ids}{A vector of specific data element IDs to retrieve. If NULL, all data elements will be retrieved.}

\item{khis_session}{The KHIS session object to use (defaults to "khis_default_session"). See `?login_to_khis()` for details.}

\item{retry}{Number of times to retry the API call in case of failure (defaults to 1).}

\item{verbosity}{Level of information to print during the API call:
- 0: No output
- 1: Show headers
- 2: Show headers and bodies
- 3: Show headers, bodies, and curl status message}
}
\value{
A tibble containing the following columns:
\describe{
 \item{element_id}{The unique identifier for the data element.}
 \item{element}{The name of the data element.}
}
}
\description{
\code{get_data_elements()} retrieves a list of data elements from the KHIS API
  server, including their IDs and names.
}
\seealso{
[login_to_khis()]
}
