% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khis_cred.R
\name{khis_cred}
\alias{khis_cred}
\title{Sets the KHIS Credentials}
\usage{
khis_cred(config_path = NULL, username = NULL, password = NULL)
}
\arguments{
\item{config_path}{An optional parameters that contains the path to configuration file with username and password}

\item{username}{The KHIS username. Can be optional if \code{config_path} is already provided}

\item{password}{The KHIS password. Can be optional if \code{config_path} is already provided or already stored using the \link{keyring} package}
}
\value{
No return value
}
\description{
Sets the KHIS Credentials
}
\details{
When \code{username} and \code{password} are provided they are persisted to the operating
system’s credential store using \link{keyring} package. Subsequent call will only
require the \code{username}. If the password has not been store a prompt for the
password will appear
}
\examples{
# These examples use the default keyring, and they are interactive,
# so, we don't run them by default
\dontrun{
  # Load username and password
  khis_cred(username = 'khis_username', password = 'PASSWORD')
}
}
\concept{cred functions}
