% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khis_cred.R
\name{khis_has_cred}
\alias{khis_has_cred}
\title{Are There Credentials on Hand?}
\usage{
khis_has_cred()
}
\value{
boolean value indicating if the credentials are available
}
\description{
Are There Credentials on Hand?
}
\examples{
khis_has_cred()
}
\seealso{
Other low-level API functions: 
\code{\link{khis_username}()}
}
\concept{low-level API functions}
