% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/khis_cred.R
\name{khis_username}
\alias{khis_username}
\title{Produces the Configured Username}
\usage{
khis_username()
}
\description{
Produces the Configured Username
}
\examples{
khis_username()
}
\seealso{
Other low-level API functions: 
\code{\link{khis_has_cred}()}
}
\concept{low-level API functions}
