% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysdata.R
\docType{data}
\name{population_data}
\alias{population_data}
\title{Population Data from KNBS Census}
\format{
A tibble with 54 rows and 4 variables:
\itemize{
\item county     - Character vector of county names.
\item subcounty  - Character vector of subcounty names.
\item age        - The age distribution.
\item sex        - The sex of the population.
\item population - The number of this population.
}
}
\source{
Kenya National Bureau of Statistics (KNBS), 2019 Kenya Population and Housing Census Volume III: Distribution of populayion by Age, Sex and Administtrative units
\url{https://www.knbs.or.ke/download/2019-kenya-population-and-housing-census-volume-iii-distribution-of-population-by-age-sex-and-administrative-units/}
}
\usage{
population_data
}
\description{
Provides population statistics for counties and subcounties in Kenya, as per Volume 3 of the 2019 Kenya Population and Housing Census Table 2.3
}
\keyword{internal}
