% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/population_data.R
\docType{data}
\name{population_data}
\alias{population_data}
\title{Population Data from KNBS Census}
\format{
A tibble with 54 rows and 4 variables:
\describe{
\item{county}{Character vector of county names.}
\item{subcounty}{Character vector of subcounty names.}
\item{population}{Numeric vector of total population counts.}
\item{...}{(List other variables and their descriptions.)}
}
}
\source{
Kenya National Bureau of Statistics (KNBS), Volume 3 of the Population Census
\url{https://www.knbs.or.ke/?page_id=1546}
}
\usage{
population_data
}
\description{
Provides population statistics for counties and subcounties in Kenya, as per Volume 3 of the KNBS population census.
}
\details{
This dataset includes:
\itemize{
\item County names
\item Subcounty names
\item Total population counts
\item ... (List other relevant variables, e.g., age groups, gender, etc.)
}
}
\keyword{internal}
